backend/pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "hr-db"
version = "0.0.1"
description = "HR DB"
authors = ["Dimitri Merejkowsky <dimitri@dmerej.info>"]
readme = "README.md"
license = "BSD-3-Clause"
repository = ""


[tool.poetry.dependencies]
python = "^3.7"

django = "^3.2.10"
django-cors-headers = "^3.10.1"
python-dotenv = "^0.19.2"
gunicorn = "^20.1"
django-components = "^0.17"
django-hamlpy = "^1.4.4"
django-bootstrap5 = "^21.3"

[tool.poetry.dev-dependencies]
invoke = "^1.5.0"
pywatchman = "^1.4.1"

black = "21.10b0"
isort = "^5.7.0"
pytest-django = "^4.4.0"
flake8 = "^4.0.1"

pytest = "^6.2"
Faker = "^10.0.0"
playwright = "^1.17.2"
pytest-playwright = "^0.2.2"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
